requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.10
psycopg2==2.9.10
redis==5.2.1
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
